QMacStyle: fix tab widget rendering
This patch-set combines two recent patches for Big Sur (hiding 'disclose'
button on a NSPopUpButton) and adjusting for (now) visible button's round
corners. Since it makes little sense to only partially fix the thing, we
also backport another change (not so recent, missed 5.12 in the past) that
corrects clipping (where we are trying to cut NSBox'es frame, which otherwise
is semi-visible through semi-transparent buttons in the 'Dark' theme).

Task-number: QTBUG-86513
Change-Id: Ieea5a47ba3073eb564594039895cf70a0a34e7be
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>